Triangle Explosion Game :D
Triangle Explosion game, when you click a triangle, it explodes and an additional 1 second is added to your 30 second timer. in order to use this, get notepad from computer, paste the code. click file, click save, put it in desktop and name it beadity.html it must have the type html or else it wont work.