BetaTriangleExplosionGame.html
this is a beta triangle explosion game, instead of a timer, it adds Points instead. in order to play, you need to make a notepad in the computer, paste the code, click File, click Save, put it on desktop and name it bead.html it must have the type .html or else you wont be able to play it.